python_ai_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
certifi==2024.2.2
openai==1.3.0
anthropic==0.7.0
numpy==1.26.2
pydantic==2.5.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
pyyaml==6.0.1
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

